NASABASE/backend/requirements.txt
flask
flask-cors
openai
chromadb
python-dotenv
